+++ boa/core/render.py
@@ -88,35 +88,45 @@ def ensure_list(x):
 def normalize_recipe(ydoc):
     # normalizing recipe:
     # sources -> list
-    # every output -> to outputs list
+    # every output -> to steps list
     if ydoc.get("context"):
         del ydoc["context"]
 
     if ydoc.get("source"):
         ydoc["source"] = ensure_list(ydoc["source"])
 
-    if not ydoc.get("outputs"):
-        ydoc["outputs"] = [{"package": ydoc["package"]}]
+    toplevel_output = None
 
-        toplevel_output = ydoc["outputs"][0]
+    if ydoc.get("outputs"):
+        ydoc["steps"] = ydoc["outputs"]
+        del ydoc["outputs"]
+
+    if not ydoc.get("steps"):
+        ydoc["steps"] = [{"package": ydoc["package"]}]
+        toplevel_output = ydoc["steps"][0]
     else:
-        for o in ydoc["outputs"]:
-            if o["package"]["name"] == ydoc["package"]["name"]:
+        for o in ydoc["steps"]:
+            if "package" not in o:
+                continue
+            if not toplevel_output and o["package"]["name"] == ydoc["package"]["name"]:
                 toplevel_output = o
-                break
-        else:
-            # how do we handle no-output toplevel?!
-            toplevel_output = None
+
+            # merge version into steps if they don't have one
+            if "version" not in o["package"]:
+                o["package"]["version"] = ydoc["package"]["version"]
+
+        # how do we handle no-output toplevel?!
+        if toplevel_output is None:
             assert not ydoc.get("requirements")
 
+    # move these under toplevel output
     if ydoc.get("requirements"):
-        # move these under toplevel output
         assert not toplevel_output.get("requirements")
         toplevel_output["requirements"] = ydoc["requirements"]
         del ydoc["requirements"]
 
+    # move these under toplevel output
     if ydoc.get("test"):
-        # move these under toplevel output
         assert not toplevel_output.get("test")
         toplevel_output["test"] = ydoc["test"]
         del ydoc["test"]
@@ -130,7 +140,6 @@ def normalize_recipe(ydoc):
 
     move_to_toplevel("run_exports")
     move_to_toplevel("ignore_run_exports")
-
     return ydoc
 
 
@@ -179,6 +188,7 @@ def render(recipe_path, config=None):
 
     flatten_selectors(ydoc, ns_cfg(config))
 
+    # Normalize the entire recipe
     ydoc = normalize_recipe(ydoc)
     # console.print("\n[yellow]Normalized recipe[/yellow]\n")
     # console.print(ydoc)