+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+0.10.0 (March 18, 2022)
+=======================
+
+- [boa] add `boa --version`
+- [boa] add more docs and vastly improved new recipe schema, render recipe schema in docs
+- [boa] add version from top-level to outputs to make validation pass
+- [boa] move CondaBuildSpec class to it's own file
+- [boa] save properly rendered recipe into final package
+- [boa] implement build steps and variant inheritance logic
+- [boa] read and respect binary_relocation value (thanks @frmdstryr)
+- [boa] add debug assert messages (thanks @dhirschfeld)
+
+
 0.9.0 (February 11, 2022)
 =========================